+++ app/server/index.js
@@ -9,6 +9,7 @@ import itemRouter from './routes/item.routes.js';
 import userRouter from './routes/user.routes.js';
 import transactionsRouter from './routes/transactions.js';
 import cartRouter from './routes/cart.routes.js';
+import discountRouter from './routes/discount.routes.js'
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -25,6 +26,7 @@ app.use('/api/items', itemRouter);
 app.use('/api/users', userRouter);  
 app.use('/api/transactions', transactionsRouter);   
 app.use('/api/cart', cartRouter);
+app.use('/api/discount', discountRouter)
 
 app.get('/', (req, res) => {
   res.send('🍿 Snack-Overflow API is up and running');