+++ sql/schema.sql
@@ -53,6 +53,8 @@ CREATE TABLE transactions(
   creation_date DATETIME NOT NULL,
   delivered_date DATETIME,
   expected_date DATETIME,
+  order_status VARCHAR(9) NOT NULL,
+  total DECIMAL(10, 2) NOT NULL,
   FOREIGN KEY (user_id) REFERENCES users(id)
     ON DELETE CASCADE
     ON UPDATE CASCADE
@@ -70,3 +72,9 @@ CREATE TABLE transaction_items(
     ON DELETE CASCADE 
     ON UPDATE CASCADE
 );
+
+CREATE TABLE discounts(
+  code VARCHAR(10) PRIMARY KEY,
+  sku CHAR(5) NOT NULL,
+  discount INT NOT NULL
+)