+++ CHANGELOG.md
@@ -42,4 +42,14 @@
 
 **V07.22**
   - Added some try/catch and more error messages handling unknown errors.
-  - Added "Already have an account" and "Don't have an account" to sign up and log in screens respectively .
+  - Added "Already have an account" and "Don't have an account" to sign up and log in screens respectively.
+  - Added discount functionality, order types, and total tracking to transaction creation.
+
+**V07.23**
+  - Added home screen.
+  - Added grocery pages (Ability to search for is still being implemented).
+  - Minor aesthetic changes to some parts of app.
+
+**V07.24**
+  - Added pattern checking for signup screen.
+  - Updated full transaction backend to include a status.